scale the summary terms by the number of documents

+++ lumongo-cluster/src/main/java/org/lumongo/server/index/analysis/TermFreq.java
@@ -48,7 +48,7 @@ public class TermFreq {
 		if (Lumongo.AnalysisRequest.TermSort.TFIDF.equals(termSort)) {
 			if (docFreq != null) {
 				for (Lumongo.Term.Builder term : terms) {
-					double score = docFreq.getScoreForTerm(term.getTermFreq() / docCount, term.getDocFreq());
+					double score = docFreq.getScoreForTerm((long) (term.getTermFreq() / (double) docCount), term.getDocFreq());
 					term.setScore(score);
 				}
 			}
